# src/main/java/healthSafe/dvds20222cg4hce/controller/request/grupofamiliar/GrupoFamiliarCompartirRequest.java
package healthSafe.dvds20222cg4hce.controller.request.grupofamiliar;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class GrupoFamiliarCompartirRequest {
	private String mail;
	private String codigo;
}