# src/main/java/healthSafe/dvds20222cg4hce/controller/request/grupofamiliar/GrupoFamiliarRequest.java
package healthSafe.dvds20222cg4hce.controller.request.grupofamiliar;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class GrupoFamiliarRequest {
	private String nombre;
}