# src/main/resources/requests/Historia-Medica_Swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/gruposFamiliares:
    post:
      tags:
      - grupo-familiar-controller-rest
      operationId: createGrupoFamiliar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrupoFamiliarRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/gruposFamiliares/{id}:
    get:
      tags:
      - grupo-familiar-controller-rest
      operationId: getGrupoFamilir
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/gruposFamiliares/compartirCodigo:
    get:
      tags:
      - grupo-familiar-controller-rest
      operationId: shareCodigo
      parameters:
      - name: compartirData
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/GrupoFamiliarCompartirRequest'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    GrupoFamiliarRequest:
      type: object
      properties:
        nombre:
          type: string
        historiaMedicaId:
          type: integer
          format: int64
...
    GrupoFamiliarCompartirRequest:
      type: object
      properties:
        mail:
          type: string
        codigo:
          type: string
...